+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Adding scraping selectors `prev_sibling`, `next_sibling`, `find_ancestor` & `last`.
 * Adding `xan scrape images`.
 * Adding `xan search -u/--url-prefix`.
+* Adding `xan network -f nodelist`.
 
 *Fixes*